Desktop/Uni/Proyectos/1º/1ev/IP/MartinezMorenoLuis -S3L3/arbolSesion2NoPresencial/TreeTest.java


import static org.junit.Assert.*;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;

/**
 * The test class TreeTest.
 *
 * @author  (your name)
 * @version (a version number or a date)
 */
public class TreeTest
{
    /**
     * Default constructor for test class TreeTest
     */
    public TreeTest()
    {
    }

    /**
     * Sets up the test fixture.
     *
     * Called before every test case method.
     */
    @Before
    public void setUp()
    {
    }

    /**
     * Tears down the test fixture.
     *
     * Called after every test case method.
     */
    @After
    public void tearDown()
    {
    }

    @Test
    public void TestTree()
    {
        Tree tree1 = new Tree();
        assertEquals(0, tree1.harvest());
        tree1.setNumberOfFruits(4);
        assertEquals(0, tree1.harvest());
        tree1.setNumberOfFruits(7);
    }

    @Test
    public void TreeTest()
    {
    }
}


